requirements.txt
altgraph==0.17.4
bidict==0.23.1
blinker==1.9.0
chardet==5.2.0
click==8.3.0
colorama==0.4.6
distlib==0.4.0
et_xmlfile==2.0.0
filelock==3.19.1
Flask==3.1.2
Flask-Login==0.6.3
Flask-SocketIO==5.3.6
h11==0.16.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
numpy==2.3.3
openpyxl==3.1.2
packaging==25.0
pandas==2.3.2
pdfkit==1.0.0
pefile==2023.2.7
pillow==10.2.0
platformdirs==4.4.0
psycopg2-binary==2.9.10
psycopg2-pool==1.2
pyinstaller==6.16.0
pyinstaller-hooks-contrib==2025.9
pymodbus==3.11.3
pypng==0.20220715.0
pyserial==3.5
python-barcode==0.15.1
python-dateutil==2.9.0.post0
python-engineio==4.9.0
python-socketio==5.11.1
pytz==2025.2
pywin32-ctypes==0.2.3
qrcode==7.4.2
redis==7.1.0
reportlab==4.1.0
setuptools==80.9.0
simple-websocket==1.1.0
six==1.17.0
ttkbootstrap==1.14.2
typing_extensions==4.15.0
tzdata==2025.2
Unidecode==1.4.0
virtualenv==20.34.0
Werkzeug==3.1.3
wsproto==1.2.0
